requirements.txt
azure-search-documents
azure-identity
semantic-kernel
langchain
langchain-openai
pandas